Mensagem Sim Não quando excluir/cancelar

+++ LunaMia7/src/controller/BotoesAcoesController.java
@@ -9,6 +9,7 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 
+import javax.swing.JOptionPane;
 import javax.swing.JScrollPane;
 import javax.swing.SwingUtilities;
 
@@ -27,6 +28,7 @@ import view.CadastroMateriaPrimaEstoque;
 import view.CriarOrcamento;
 import view.MateriaPrimaView;
 import view.Mensagem;
+import view.MensagemSimNao;
 import view.Orcamentos;
 import view.VisualizarMateriaPrima;
 import view.VisualizarOrcamento;
@@ -211,12 +213,26 @@ public class BotoesAcoesController extends ComponentAdapter {
 	// MÉTODO EXCLUIR MATERIA PRIMA
 	public void excluirMP(MateriaPrima MP) {
 
-		int linhaSelecionada = materiaPrimaView.getTabelaMateriaPrima().getSelectedRow();
+	    boolean emUso = orcamProdDAO.existePorIdMateriaPrima(MP.getIdMateriaPrima());
 
-		materiaPrimaView.tabelaModeloMateriaPrima.removerMatPrima(linhaSelecionada);
-		materiaPrimaDAO.excluirMateriasPrimas(MP);
+	    if (emUso) {
+	        Mensagem.mostrar(null, "Atenção", "Essa matéria prima está \nvinculada a um ou mais orçamentos \ne não pode ser excluída!");
+	        return;
+	    }
 
-		Mensagem.mostrar(null, "Informação", "Materia prima exluída com sucesso!");
+	    int linhaSelecionada = materiaPrimaView.getTabelaMateriaPrima().getSelectedRow();
+	   
+		int confirm = MensagemSimNao.mostrar(null, "Excluir", "Deseja excluir a matéria prima? \nEsta ação não poderá ser desfeita.");
+
+		if (confirm == JOptionPane.YES_OPTION) {
+			
+		    materiaPrimaView.tabelaModeloMateriaPrima.removerMatPrima(linhaSelecionada);
+		    materiaPrimaDAO.excluirMateriasPrimas(MP);
+
+		    Mensagem.mostrar(null, "Informação", "Materia prima excluída com sucesso!");
+
+		}
+
 	}
 
 	// RECARREGANDO TABELA MATERIA PRIMA
@@ -422,6 +438,7 @@ public class BotoesAcoesController extends ComponentAdapter {
 
 			if (confirOrcam != null) {
 				confirOrcam.setOrcamento(OrcEditada);
+				
 				confirOrcam.setFormPagamento(criarOrcamento.getCbFormaPaga().getSelectedItem().toString());
 				
 				if(!dataConfirStr.trim().isEmpty()) {
@@ -474,14 +491,20 @@ public class BotoesAcoesController extends ComponentAdapter {
 		
 		int linhaSelecionada = orcamentos.getTabelaOrcamentos().getSelectedRow();
 		
-		orcamProdDAO.excluirPorIdOrcamento(orcamento.getIdOrcamento());
-		confirOrcamDAO.excluirConfirOrcam(orcamento.getIdOrcamento());
 		
-		orcamentos.tabelaModeloOrcamento.removerOrcamento(linhaSelecionada);
-		orcamentoDAO.excluirOrcamento(orcamento);
-		clienteDAO.excluirCliente(orcamento.getCliente().getIdCliente());
-		
-		Mensagem.mostrar(null, "Informação", "Orçamento exluído com sucesso!");
+		int confirm = MensagemSimNao.mostrar(null, "Excluir", "Deseja excluir o orçamento? \nEsta ação não poderá desfeita.");
+
+		if (confirm == JOptionPane.YES_OPTION) {
+			
+			orcamProdDAO.excluirPorIdOrcamento(orcamento.getIdOrcamento());
+			confirOrcamDAO.excluirConfirOrcam(orcamento.getIdOrcamento());
+			
+			orcamentos.tabelaModeloOrcamento.removerOrcamento(linhaSelecionada);
+			orcamentoDAO.excluirOrcamento(orcamento);
+			clienteDAO.excluirCliente(orcamento.getCliente().getIdCliente());
+			
+			Mensagem.mostrar(null, "Informação", "Orçamento exluído com sucesso!");
+		}
 		
 	}
 	
@@ -496,6 +519,10 @@ public class BotoesAcoesController extends ComponentAdapter {
 		
 		ConfirOrcamDAO confirOrcamDAO = new ConfirOrcamDAO();
 		Cliente cliente = orcamento.getCliente();
+		
+		int confirm = MensagemSimNao.mostrar(null, "Cancelar", "Deseja cancelar o orçamento? \nEsta ação não poderá desfeita, \ne a matéria prima voltará para o estoque.");
+
+		if (confirm == JOptionPane.YES_OPTION) {
 		
 		List<OrcamentoProduto> listaProdUtiORC = orcamProdDAO.listarOrcamProd();
 		List<MateriaPrima> listaMP = materiaPrimaDAO.listarMateriaPrima(usuarioLogado.getEmail());
@@ -529,7 +556,7 @@ public class BotoesAcoesController extends ComponentAdapter {
 		clienteDAO.excluirCliente(orcamento.getCliente().getIdCliente());
 		
 		Mensagem.mostrar(null, "Informação", "Orçamento cancelado com sucesso! \nQuantidade das materias primas \nutilizadas no seu orçamento \natualizadas com sucesso! ");
-		
+		}
 	}
 	
 	public void atualizarQuantMP() {
